Move misplaced join overloads to fmt/ranges.h
